medical/ai-voice-response-system/requirements.txt
# 音声認識
openai-whisper>=20231117

# 音声合成（複数のオプション）
# オプション1: pyttsx3（ローカル・無料・簡単）
pyttsx3>=2.90

# オプション2: transformers TTS（高品質）
transformers>=4.35.0
torch>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0

# Webサーバー
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# AI応答生成（オプション）
openai>=1.3.0

# その他
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
